// Front/ProEventos-App/src/app/models/Evento.ts
import { RedeSocial } from './RedeSocial';
import { Lote } from "./Lote";
// import { Palestrante } from './Palestrante';

export interface Evento {
  id: number;
  local: string;
  dataEvento?: string;
  tema: string;
  qtdPessoas: number;
  imagemURL: string;
  telefone: string;
  email: string;
  lotes: Lote[];
  redesSociais: RedeSocial[];
//   palestrantesEventos: Palestrante[];
}